+++ src/type_set.rs
@@ -1,9 +1,13 @@
 use amethyst::ecs::prelude::*;
 use amethyst::ecs::shred::Resource;
+use crossbeam_channel;
 use serde::Serialize;
+use serde::de::DeserializeOwned;
 use std::marker::PhantomData;
 
-use {EditorConnection, SyncComponentSystem, SyncResourceSystem};
+use {EditorConnection, DeserializerMap, SyncComponentSystem};
+use read_resource::ReadResourceSystem;
+use write_resource::WriteResourceSystem;
 
 /// Create a set of types, where the value is the stringified typename.
 #[macro_export]
@@ -77,12 +81,20 @@ where
     T: ResourceSet,
 {
     /// Create a resource-synchronization system for each type in the set.
-    pub(crate) fn create_resource_sync_systems(
+    pub(crate) fn create_resource_read_systems(
         &self,
         dispatcher: &mut DispatcherBuilder,
         connection: &EditorConnection,
     ) {
-        T::create_sync_systems(dispatcher, connection, &self.names);
+        T::create_read_systems(dispatcher, connection, &self.names);
+    }
+
+    pub(crate) fn create_resource_write_systems(
+        &self,
+        dispatcher: &mut DispatcherBuilder,
+        deserializer_map: &mut DeserializerMap,
+    ) {
+        T::create_write_systems(dispatcher, deserializer_map, &self.names);
     }
 }
 
@@ -152,17 +164,31 @@ pub trait ResourceSet {
     ///
     /// Their names are passed in the order they are inserted into the type set.
     /// Returns the number of systems created.
-    fn create_sync_systems(
+    fn create_read_systems(
         dispatcher: &mut DispatcherBuilder,
         connection: &EditorConnection,
+        names: &[&'static str],
+    ) -> usize;
+
+    fn create_write_systems(
+        dispatcher: &mut DispatcherBuilder,
+        deserializer_map: &mut DeserializerMap,
         names: &[&'static str],
     ) -> usize;
 }
 
 impl ResourceSet for () {
-    fn create_sync_systems(
+    fn create_read_systems(
         _: &mut DispatcherBuilder,
         _: &EditorConnection,
+        _: &[&'static str],
+    ) -> usize {
+        0
+    }
+
+    fn create_write_systems(
+        _: &mut DispatcherBuilder,
+        _: &mut DeserializerMap,
         _: &[&'static str],
     ) -> usize {
         0
@@ -171,18 +197,33 @@ impl ResourceSet for () {
 
 impl<T> ResourceSet for (T,)
 where
-    T: Resource + Serialize,
+    T: Resource + Serialize + DeserializeOwned,
 {
-    fn create_sync_systems(
+    fn create_read_systems(
         dispatcher: &mut DispatcherBuilder,
         connection: &EditorConnection,
         names: &[&'static str],
     ) -> usize {
         dispatcher.add(
-            SyncResourceSystem::<T>::new(names[0], connection.clone()),
+            ReadResourceSystem::<T>::new(names[0], connection.clone()),
             "",
             &[],
         );
+        1
+    }
+
+    fn create_write_systems(
+        dispatcher: &mut DispatcherBuilder,
+        deserializer_map: &mut DeserializerMap,
+        names: &[&'static str],
+    ) -> usize {
+        let (sender, receiver) = crossbeam_channel::unbounded();
+        dispatcher.add(
+            WriteResourceSystem::<T>::new(names[0], receiver),
+            "",
+            &[],
+        );
+        deserializer_map.insert(names[0], sender);
         1
     }
 }
@@ -192,12 +233,21 @@ where
     T: ResourceSet,
     U: ResourceSet,
 {
-    fn create_sync_systems(
+    fn create_read_systems(
         dispatcher: &mut DispatcherBuilder,
         connection: &EditorConnection,
         names: &[&'static str],
     ) -> usize {
-        let idx = T::create_sync_systems(dispatcher, connection, names);
-        idx + U::create_sync_systems(dispatcher, connection, &names[idx..])
+        let idx = T::create_read_systems(dispatcher, connection, names);
+        idx + U::create_read_systems(dispatcher, connection, &names[idx..])
+    }
+
+    fn create_write_systems(
+        dispatcher: &mut DispatcherBuilder,
+        deserializer_map: &mut DeserializerMap,
+        names: &[&'static str],
+    ) -> usize {
+        let idx = T::create_write_systems(dispatcher, deserializer_map, names);
+        idx + U::create_write_systems(dispatcher, deserializer_map, &names[idx..])
     }
 }